+++ mne/viz/tests/test_3d.py
@@ -15,18 +15,16 @@ import numpy as np
 from numpy.testing import assert_raises, assert_equal
 
 from mne import (make_field_map, pick_channels_evoked, read_evokeds,
-                 read_trans, read_dipole, SourceEstimate, make_sphere_model,
-                 read_source_estimate)
+                 read_trans, read_dipole, SourceEstimate, make_sphere_model)
 from mne.io import read_raw_ctf, read_raw_bti, read_raw_kit, read_info
 from mne.io.meas_info import write_dig
 from mne.viz import (plot_sparse_source_estimates, plot_source_estimates,
                      plot_trans, snapshot_brain_montage, plot_head_positions,
-                     plot_alignment, plot_glass_brain)
+                     plot_alignment)
 from mne.viz.utils import _fake_click
 from mne.utils import (requires_mayavi, requires_pysurfer, run_tests_if_main,
-                       _import_mlab, _TempDir, requires_nibabel, check_version,
-                       requires_nilearn)
-from mne.datasets import testing, spm_face
+                       _import_mlab, _TempDir, requires_nibabel, check_version)
+from mne.datasets import testing
 from mne.source_space import read_source_spaces
 from mne.bem import read_bem_solution, read_bem_surfaces
 
@@ -366,32 +364,6 @@ def test_snapshot_brain_montage():
 
     # Make sure we raise error if the figure has no scene
     assert_raises(TypeError, snapshot_brain_montage, fig, info)
-
-
-@requires_nilearn
-@spm_face.requires_spm_data
-@testing.requires_testing_data
-@requires_nibabel()
-def test_plot_glass_brain():
-    """Test plotting stc with nilearn."""
-    import matplotlib.pyplot as plt
-    subj_dir = op.join(spm_face.data_path(), 'subjects')
-    src_fname = op.join(spm_face.data_path(), 'subjects', 'spm', 'bem',
-                        'spm-oct-6-src.fif')
-    sample_src = read_source_spaces(src_fname)
-    vertices = [s['vertno'] for s in sample_src]
-    n_time = 5
-    n_verts = sum(len(v) for v in vertices)
-    stc_data = np.ones((n_verts * n_time))
-    stc_data.shape = (n_verts, n_time)
-    stc = SourceEstimate(stc_data, vertices, 1, 1, 'spm')
-    # spm dataset has the ribbon.mgz.
-    plot_glass_brain(stc, 'spm', subj_dir, src=None, initial_time=0.)
-
-    stc = read_source_estimate(fname_vol)  # Volume stc.
-    src = read_source_spaces(fwd_fname)
-    plot_glass_brain(stc, 'sample', subjects_dir, src=src)
-    plt.close('all')
 
 
 run_tests_if_main()